refactor: move services to providers
