+++ studio/src/app/modals/editor/app-custom-data/app-custom-data.tsx
@@ -4,8 +4,8 @@ import i18n from '../../../stores/i18n.store';
 
 import {Constants} from '../../../types/core/constants';
 
-import {getStorageService, StorageService} from '../../../services/storage/storage.service';
-import {StorageOfflineService} from '../../../services/storage/storage.offline.service';
+import {getStorageService, StorageService} from '../../../providers/storage/storage.service';
+import {StorageOfflineService} from '../../../providers/storage/storage.offline.service';
 
 import {AppIcon} from '../../../components/core/app-icon/app-icon';
